+++ ni/src/ncl/VarSupport.c
@@ -1350,11 +1350,11 @@ struct _NclVarRec* self;
 			if(ret < 0) {
 				return(NhlWARNING);
 			}
-			ret = nclfprintf(fp,"Total Size: %lld bytes\n",thevalue->multidval.totalsize);
+			ret = nclfprintf(fp,"Total Size: %lld bytes\n",(long long)thevalue->multidval.totalsize);
 			if(ret < 0) {
 				return(NhlWARNING);
 			}
-			ret = nclfprintf(fp,"            %lld values\n",thevalue->multidval.totalelements);
+			ret = nclfprintf(fp,"            %lld values\n",(long long)thevalue->multidval.totalelements);
 			if(ret < 0) {
 				return(NhlWARNING);
 			}
@@ -1377,7 +1377,7 @@ struct _NclVarRec* self;
 						return(NhlWARNING);
 					}
 				}
-				ret = nclfprintf(fp,"%lld]",self->var.dim_info[i].dim_size);
+				ret = nclfprintf(fp,"%lld]",(long long)self->var.dim_info[i].dim_size);
 				if(ret < 0) {
 					return(NhlWARNING);
 				}
